+++ src/compose/PosterOptions/TagSelector.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useCallback, useRef } from 'react';
-import { View, StyleSheet, ScrollView, TouchableOpacity } from 'react-native';
+import React, { FC, useCallback, useEffect, useRef } from 'react';
+import { ScrollView, StyleSheet, TouchableOpacity, View } from 'react-native';
 import { observer, useLocalStore } from 'mobx-react';
 import MIcon from 'react-native-vector-icons/MaterialCommunityIcons';
 
@@ -11,6 +11,10 @@ import hashtagService from '../../common/services/hashtag.service';
 import HistoryStore from '../../common/stores/HistoryStore';
 import TextInput from '../../common/components/TextInput';
 import MText from '../../common/components/MText';
+import { StackScreenProps } from '@react-navigation/stack';
+import { PosterStackParamList } from '~/compose/PosterOptions/PosterStackNavigator';
+import { useComposeContext } from '~/compose/useComposeStore';
+import { BottomSheetScrollView } from '@gorhom/bottom-sheet';
 
 /**
  * Tag row
@@ -38,12 +42,16 @@ const TagRow = props => {
   );
 };
 
+interface TagSelectorProps
+  extends FC,
+    StackScreenProps<PosterStackParamList, 'TagSelector'> {}
+
 /**
  * Tag selector
  */
-export default observer(function (props) {
+export default observer(function ({}: TagSelectorProps) {
   const theme = ThemedStyles.style;
-  const store = props.route?.params?.store;
+  const store = useComposeContext();
   const inputRef = useRef<any>(null);
 
   const localStore = useLocalStore(
@@ -117,7 +125,8 @@ export default observer(function (props) {
       </MText>
       <View style={styles.suggestedContainer}>
         <ScrollView
-          showsHorizontalScrollIndicator={false}
+          style={theme.flexContainer}
+          showsHorizontalScrollIndicator={true}
           contentContainerStyle={styles.suggestedScroll}
           horizontal={true}>
           <MIcon name="fire" size={23} style={theme.colorAlert} />
@@ -172,7 +181,7 @@ export default observer(function (props) {
               Clear history
             </MText>
           </View>
-          <ScrollView keyboardShouldPersistTaps={'handled'}>
+          <BottomSheetScrollView keyboardShouldPersistTaps={'handled'}>
             {localStore.history.history.map(t => (
               <MText
                 style={styles.historyTag}
@@ -180,7 +189,7 @@ export default observer(function (props) {
                 #{t}
               </MText>
             ))}
-          </ScrollView>
+          </BottomSheetScrollView>
         </View>
       )}
       <View style={styles.tagsContainer}>
@@ -195,6 +204,7 @@ export default observer(function (props) {
 const styles = StyleSheet.create({
   suggestedContainer: {
     height: 50,
+    width: '100%',
   },
   historyTag: {
     fontSize: 17,